microservices/admin-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
cassandra-driver==3.28.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
pydantic==2.5.3
